pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "datasetx"
version = "0.0.1"
description = "Commonly used datasets based on torch and torchvision"
authors = [
  { name="xyfJASON", email="xyfjason@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "torch>=2.4.1",
    "torchvision>=0.19.1",
    "torchaudio>=2.4.1",
    "numpy",
    "pandas",
    "Pillow",
    "h5py",
    "opencv-python",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/xyfJASON/datasetx"
Issues = "https://github.com/xyfJASON/datasetx/issues"
